Start to accept urls
